remove files + move Heat.tsx into /web
