--- src/web/components/loginButton.tsx
@@ -1,9 +1,0 @@
-import { useAuth0 } from "@auth0/auth0-react";
-
-const LoginButton = () => {
-    const { loginWithRedirect } = useAuth0();
-
-    return <button className={"btn btn-primary"} onClick={() => loginWithRedirect()}>Log In</button>;
-};
-
-export default LoginButton;
