--- src/web/components/logoutButton.tsx
@@ -1,13 +1,0 @@
-import { useAuth0 } from "@auth0/auth0-react";
-
-const LogoutButton = () => {
-    const { logout } = useAuth0();
-
-    return (
-        <button className={"btn btn-outline btn-error"} onClick={() => logout({ logoutParams: { returnTo: window.location.origin } })}>
-            Log Out
-        </button>
-    );
-};
-
-export default LogoutButton;
