+++ src/web/main.tsx
@@ -2,7 +2,7 @@ import {createRoot} from 'react-dom/client'
 import Home from "./home.tsx";
 import {StrictMode} from "react";
 import {BrowserRouter, Route, Routes} from "react-router";
-import Heat from "./components/Heat.tsx"
+import Heat from "./Heat.tsx"
 import Loading from "./loading.tsx";
 
 export default function App() {
